pyproject.toml
[project]
name = "20251206-music-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "discord-py[voice]>=2.6.4",
    "fakeredis>=2.32.1",
    "httpx>=0.28.1",
    "montydb>=2.5.6",
    "orjson>=3.11.4",
    "python-dotenv>=1.2.1",
    "pytubefix>=10.3.5",
    "scrapetube>=2.6.0",
    "yt-dlp>=2025.11.12",
]
